refactor(v010): carve voicegw logs out of _legacy
Seventh per-command carve-out in section 2.

Adds voicegateway/cli/logs.py owning the 'logs' command. Body
unchanged from v0.0.5; the lazy 'import datetime' that lived inside
the function body is hoisted to the top of the new module.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (8 source
files now).
