+++ voicegateway/cli/_legacy.py
@@ -50,51 +50,6 @@ def _load_gateway(config_path: str | None):
         raise typer.Exit(1) from e
 
 
-@app.command(name="logs")
-def logs_cmd(
-    config: str = typer.Option(None, "--config", "-c", help="Path to voicegw.yaml"),
-    project: str = typer.Option(None, "--project", "-p", help="Filter by project"),
-    tail: int = typer.Option(20, "--tail", "-n", help="Number of rows"),
-    modality: str = typer.Option(None, "--modality", "-m", help="stt, llm, or tts"),
-):
-    """Show recent request logs."""
-    gw = _load_gateway(config)
-    if gw.storage is None:
-        console.print("[yellow]Cost tracking is not enabled in voicegw.yaml[/yellow]")
-        raise typer.Exit(0)
-
-    rows = asyncio.run(
-        gw.storage.get_recent_requests(limit=tail, modality=modality, project=project)
-    )
-    if not rows:
-        console.print("[dim]No logs found.[/dim]")
-        return
-
-    table = Table(title=f"Recent Requests ({len(rows)})")
-    table.add_column("Time", style="cyan")
-    table.add_column("Project", style="magenta")
-    table.add_column("Modality")
-    table.add_column("Model")
-    table.add_column("Cost", justify="right")
-    table.add_column("Latency", justify="right")
-    table.add_column("Status")
-
-    import datetime
-
-    for r in rows:
-        ts = datetime.datetime.fromtimestamp(r["timestamp"]).strftime("%H:%M:%S")
-        table.add_row(
-            ts,
-            r.get("project") or "-",
-            r.get("modality", "").upper(),
-            r.get("model_id", ""),
-            f"${r.get('cost_usd', 0):.6f}",
-            f"{int(r.get('total_latency_ms') or 0)}ms",
-            r.get("status", ""),
-        )
-    console.print(table)
-
-
 @app.command(name="smoke-test")
 def smoke_test(
     config: str = typer.Option(None, "--config", "-c", help="Path to voicegw.yaml"),
